Add top-level creation URL for global scope
Global scopes have two creation URLs: one for itself and
one for the "top-level" scope. It's not immediately
obvious what is considered top-level here (it is not strictly
defined in the specification).

In any case, reports need the creation URL of the scope itself,
not the top-level version. Therefore, propagate this information
from all scopes, where the worker and worklets remain to pass
in `None` for their top-level scope.

Part of #37328

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>